+++ rust/Cargo.toml
@@ -4,4 +4,4 @@
 # artifacts, speeding up compilation considerably.
 
 [workspace]
-members = ["pact_matching", "pact_mock_server", "pact_mock_server_async", "pact_mock_server_cli", "pact_consumer", "pact_verifier", "pact_verifier_cli"]
+members = ["pact_matching", "pact_mock_server", "pact_mock_server_cli", "pact_consumer", "pact_verifier", "pact_verifier_cli"]
