+++ rust/pact_mock_server/src/lib.rs
@@ -47,44 +47,33 @@
 
 #![warn(missing_docs)]
 
-#[macro_use] extern crate log;
-#[allow(unused_imports)] #[macro_use] extern crate p_macro;
-#[macro_use] extern crate lazy_static;
-extern crate libc;
-#[macro_use] extern crate pact_matching;
-#[macro_use] extern crate serde_json;
-extern crate env_logger;
+#[allow(unused_imports)] extern crate p_macro;
+extern crate pact_matching;
+extern crate serde_json;
 extern crate hyper;
+extern crate futures;
+extern crate tokio;
+#[macro_use] extern crate log;
 extern crate uuid;
 extern crate itertools;
+#[macro_use] extern crate lazy_static;
+extern crate libc;
 
-#[cfg(test)]
-#[macro_use]
-extern crate maplit;
+mod server;
+mod server_manager;
 
+use pact_matching::models::{Pact, Interaction, Request, PactSpecification};
+use pact_matching::Mismatch;
+use pact_matching::s;
 use libc::{c_char, int32_t};
 use std::ffi::CStr;
 use std::ffi::CString;
 use std::str;
 use std::panic::catch_unwind;
-use pact_matching::models::{Pact, Interaction, Request, OptionalBody, PactSpecification};
-use pact_matching::models::parse_query_string;
-use pact_matching::models::matchingrules::*;
-use pact_matching::models::generators::*;
-use pact_matching::Mismatch;
-use std::collections::{BTreeMap, HashMap};
-use std::thread;
 use std::sync::Mutex;
-use std::sync::mpsc::channel;
-use std::io::{self, Read, Write};
-use std::path::PathBuf;
-use hyper::server::{Server, Listening};
-use hyper::status::StatusCode;
-use hyper::header::{Headers, ContentType, AccessControlAllowOrigin, ContentLength};
-use hyper::mime::{Mime, TopLevel, SubLevel, Attr, Value};
-use hyper::uri::RequestUri;
+use serde_json::json;
 use uuid::Uuid;
-use itertools::Itertools;
+use server_manager::ServerManager;
 
 /// Enum to define a match result
 #[derive(Debug, Clone, PartialEq)]
@@ -148,228 +137,21 @@ fn mismatches_to_json(request: &Request, mismatches: &Vec<Mismatch>) -> serde_js
     })
 }
 
-/// Struct to represent a mock server
-#[derive(Debug, Clone)]
-pub struct MockServer {
-    /// Mock server unique ID
-    pub id: String,
-    /// Port the mock server is running on
-    pub port: i32,
-    /// Address of the server implementing the `Listening` trait
-    pub server: u64,
-    /// List of all match results for requests this mock server has received
-    pub matches: Vec<MatchResult>,
-    /// List of resources that need to be cleaned up when the mock server completes
-    pub resources: Vec<CString>,
-    /// Pact that this mock server is based on
-    pub pact: Pact
+/// Mock server errors
+pub enum MockServerError {
+  /// Invalid Pact Json
+  InvalidPactJson,
+  /// Failed to start the mock server
+  MockServerFailedToStart
 }
 
-impl MockServer {
-    /// Creates a new mock server with the given ID and pact
-    pub fn new(id: String, pact: &Pact) -> MockServer {
-        MockServer { id: id.clone(), port: -1, server: 0, matches: vec![], resources: vec![],
-            pact : pact.clone() }
-    }
-
-    /// Sets the port that the mock server is listening on
-    pub fn port(&mut self, port: i32) {
-        self.port = port;
-    }
-
-    /// Sets the address of the server implementing the `Listening` trait
-    pub fn server(&mut self, server: &Listening) {
-        let p = server as *const Listening;
-        self.server = p as u64;
-    }
-
-    /// Converts this mock server to a `Value` struct
-    pub fn to_json(&self) -> serde_json::Value {
-        json!({
-            "id" : json!(self.id.clone()),
-            "port" : json!(self.port as u64),
-            "provider" : json!(self.pact.provider.name.clone()),
-            "status" : json!(if self.mismatches().is_empty() { "ok" } else { "error" })
-        })
-    }
-
-    /// Returns all the mismatches that have occurred with this mock server
-    pub fn mismatches(&self) -> Vec<MatchResult> {
-        let mismatches = self.matches.iter()
-            .filter(|m| !m.matched())
-            .map(|m| m.clone());
-        let interactions: Vec<&Interaction> = self.matches.iter().map(|m| {
-            match *m {
-                MatchResult::RequestMatch(ref interaction) => Some(interaction),
-                MatchResult::RequestMismatch(ref interaction, _) => Some(interaction),
-                MatchResult::RequestNotFound(_) => None,
-                MatchResult::MissingRequest(_) => None
-            }
-        }).filter(|o| o.is_some()).map(|o| o.unwrap()).collect();
-        let missing = self.pact.interactions.iter()
-            .filter(|i| !interactions.contains(i))
-            .map(|i| MatchResult::MissingRequest(i.clone()));
-        mismatches.chain(missing).collect()
-    }
-
-    /// Mock server writes it pact out to the provided directory
-    pub fn write_pact(&self, output_path: &Option<String>) -> io::Result<()> {
-        let pact_file_name = self.pact.default_file_name();
-        let filename = match *output_path {
-            Some(ref path) => {
-                let mut path = PathBuf::from(path);
-                path.push(pact_file_name);
-                path
-            },
-            None => PathBuf::from(pact_file_name)
-        };
-        info!("Writing pact out to '{}'", filename.display());
-        match self.pact.write_pact(filename.as_path(), PactSpecification::V3) {
-            Ok(_) => Ok(()),
-            Err(err) => {
-                warn!("Failed to write pact to file - {}", err);
-                Err(err)
-            }
-        }
-    }
-
-    /// Returns the URL of the mock server
-    pub fn url(&self) -> String {
-        format!("http://localhost:{}", self.port)
-    }
-}
-
-impl PartialEq for MockServer {
-    fn eq(&self, other: &MockServer) -> bool {
-        self.id == other.id
-    }
-}
-
+///
+/// A global thread-safe, "init-on-demand" reference to a server manager.
+/// When the server manager is initialized, it starts a separate thread on which
+/// to serve requests.
+///
 lazy_static! {
-    static ref MOCK_SERVERS: Mutex<BTreeMap<String, Box<MockServer>>> = Mutex::new(BTreeMap::new());
-}
-
-fn match_request(req: &Request, interactions: &Vec<Interaction>) -> MatchResult {
-    let match_results = interactions
-        .into_iter()
-        .map(|i| (i.clone(), pact_matching::match_request(i.request.clone(), req.clone())))
-        .sorted_by(|i1, i2| {
-            let list1 = i1.1.clone().into_iter().map(|m| m.mismatch_type()).unique().count();
-            let list2 = i2.1.clone().into_iter().map(|m| m.mismatch_type()).unique().count();
-            Ord::cmp(&list1, &list2)
-        });
-    match match_results.first() {
-        Some(res) => {
-            if res.1.is_empty() {
-                MatchResult::RequestMatch(res.0.clone())
-            } else if method_or_path_mismatch(&res.1) {
-                MatchResult::RequestNotFound(req.clone())
-            } else {
-                MatchResult::RequestMismatch(res.0.clone(), res.1.clone())
-            }
-        },
-        None => MatchResult::RequestNotFound(req.clone())
-    }
-}
-
-fn method_or_path_mismatch(mismatches: &Vec<Mismatch>) -> bool {
-    let mismatch_types: Vec<String> = mismatches.iter()
-        .map(|mismatch| mismatch.mismatch_type())
-        .collect();
-    mismatch_types.contains(&s!("MethodMismatch")) || mismatch_types.contains(&s!("PathMismatch"))
-}
-
-fn extract_path(uri: &RequestUri) -> String {
-    match uri {
-        &RequestUri::AbsolutePath(ref s) => s!(s.splitn(2, "?").next().unwrap_or("/")),
-        &RequestUri::AbsoluteUri(ref url) => s!(url.path()),
-        _ => uri.to_string()
-    }
-}
-
-fn extract_query_string(uri: &RequestUri) -> Option<HashMap<String, Vec<String>>> {
-    match uri {
-        &RequestUri::AbsolutePath(ref s) => {
-            if s.contains("?") {
-                match s.splitn(2, "?").last() {
-                    Some(q) => parse_query_string(&s!(q)),
-                    None => None
-                }
-            } else {
-                None
-            }
-        },
-        &RequestUri::AbsoluteUri(ref url) => match url.query() {
-            Some(q) => parse_query_string(&s!(q)),
-            None => None
-        },
-      _ => None
-    }
-}
-
-fn extract_headers(headers: &Headers) -> Option<HashMap<String, String>> {
-    if headers.len() > 0 {
-        Some(headers.iter().map(|h| (s!(h.name()), h.value_string()) ).collect())
-    } else {
-        None
-    }
-}
-
-fn extract_body(req: &mut hyper::server::Request) -> OptionalBody {
-    let mut buffer = Vec::new();
-    match req.read_to_end(&mut buffer) {
-      Ok(size) => if size > 0 {
-        OptionalBody::Present(buffer)
-      } else {
-        OptionalBody::Empty
-      },
-      Err(err) => {
-        warn!("Failed to read request body: {}", err);
-        OptionalBody::Empty
-      }
-    }
-}
-
-fn hyper_request_to_pact_request(req: &mut hyper::server::Request) -> Request {
-    Request {
-        method: req.method.to_string(),
-        path: extract_path(&req.uri),
-        query: extract_query_string(&req.uri),
-        headers: extract_headers(&req.headers),
-        body: extract_body(req),
-        matching_rules: MatchingRules::default(),
-        generators: Generators::default()
-    }
-}
-
-fn error_body(req: &Request, error: &String) -> String {
-    let body = json!({ "error" : format!("{} : {:?}", error, req) });
-    body.to_string()
-}
-
-fn insert_new_mock_server(id: &String, pact: &Pact) {
-    MOCK_SERVERS.lock().unwrap().insert(id.clone(), Box::new(MockServer::new(id.clone(), pact)));
-}
-
-fn update_mock_server<R>(id: &String, f: &Fn(&mut MockServer) -> R) -> Option<R> {
-    match MOCK_SERVERS.lock().unwrap().get_mut(id) {
-        Some(mock_server) => Some(f(mock_server)),
-        _ => None
-    }
-}
-
-fn update_mock_server_by_port<R>(port: i32, f: &Fn(&mut MockServer) -> R) -> Option<R> {
-    let mut map = MOCK_SERVERS.lock().unwrap();
-    match map.iter_mut().find(|ms| ms.1.port == port ) {
-        Some(mock_server) => Some(f(mock_server.1)),
-        None => None
-    }
-}
-
-fn record_result(id: &String, match_result: &MatchResult) {
-    update_mock_server(id, &|mock_server: &mut MockServer| {
-        mock_server.matches.push(match_result.clone());
-    });
+    static ref MANAGER: Mutex<Option<ServerManager>> = Mutex::new(Option::None);
 }
 
 /// Starts a mock server with the given ID, pact and port number. The ID needs to be unique. A port
@@ -382,179 +164,10 @@ fn record_result(id: &String, match_result: &MatchResult) {
 ///
 /// - If a mock server is not able to be started
 pub fn start_mock_server(id: String, pact: Pact, port: i32) -> Result<i32, String> {
-    insert_new_mock_server(&id, &pact);
-    let (out_tx, out_rx) = channel();
-    let (in_tx, in_rx) = channel();
-    in_tx.send((id.clone(), pact, port)).unwrap();
-    thread::spawn(move || {
-        let (mock_server_id, pact, port) = in_rx.recv().unwrap();
-        let server = Server::http(format!("0.0.0.0:{}", port).as_str()).unwrap();
-        let server_result = server.handle(move |mut req: hyper::server::Request, mut res: hyper::server::Response| {
-            debug!("--> Hyper request to mock server {}", mock_server_id);
-            match lookup_mock_server(mock_server_id.clone(), &|_| ()) {
-                None => {
-                    warn!("Mock server {} has been shutdown", mock_server_id);
-                    *res.status_mut() = StatusCode::NotImplemented;
-                    res.headers_mut().set_raw("X-Pact", vec!["Mock server has been shut down".as_bytes().to_vec()]);
-                },
-                Some(_) => {
-                    debug!("Creating pact request from hyper request");
-                    let req = hyper_request_to_pact_request(&mut req);
-                    info!("Received request {:?}", req);
-                    let match_result = match_request(&req, &pact.interactions);
-                    record_result(&mock_server_id, &match_result);
-                    match match_result {
-                        MatchResult::RequestMatch(ref interaction) => {
-                            let response = pact_matching::generate_response(&interaction.response);
-                            info!("Request matched, sending response {:?}", response);
-                            info!("     body: '{}'\n\n", interaction.response.body.str_value());
-                            info!("     body: '{}'\n\n", interaction.response.body.str_value());
-                            *res.status_mut() = StatusCode::from_u16(response.status);
-                            res.headers_mut().set(AccessControlAllowOrigin::Any);
-                            match response.headers {
-                                Some(ref headers) => {
-                                    for (k, v) in headers.clone() {
-                                        res.headers_mut().set_raw(k, vec![v.into_bytes()]);
-                                    }
-                                },
-                                None => ()
-                            }
-                            match response.body {
-                                OptionalBody::Present(ref body) => {
-                                    res.send(body).unwrap();
-                                },
-                                _ => ()
-                            }
-                        },
-                        _ => {
-                            *res.status_mut() = StatusCode::InternalServerError;
-                            res.headers_mut().set(
-                                ContentType(Mime(TopLevel::Application, SubLevel::Json,
-                                                 vec![(Attr::Charset, Value::Utf8)]))
-                            );
-                            res.headers_mut().set(AccessControlAllowOrigin::Any);
-                            res.headers_mut().set_raw("X-Pact", vec![match_result.match_key().as_bytes().to_vec()]);
-                            let body = error_body(&req, &match_result.match_key());
-                            res.headers_mut().set(ContentLength(body.as_bytes().len() as u64));
-                            let mut res = res.start().unwrap();
-                            res.write_all(body.as_bytes()).unwrap();
-                        }
-                    }
-                }
-            }
-        });
-
-      match server_result {
-            Ok(ref server) => {
-                let port = server.socket.port() as i32;
-                info!("Mock Provider Server started on port {}", port);
-                        update_mock_server(&id, &|mock_server| {
-                    mock_server.port(port);
-                    mock_server.server(server);
-                        });
-                out_tx.send(Ok(port)).unwrap();
-            },
-            Err(e) => {
-                    error!("Could not start server: {}", e);
-                    out_tx.send(Err(format!("Could not start server: {}", e))).unwrap();
-            }
-      }
-    });
-
-    out_rx.recv().unwrap()
-}
-
-/// Looks up the mock server by ID, and passes it into the given closure. The result of the
-/// closure is returned wrapped in an `Option`. If no mock server is found with that ID, `None`
-/// is returned.
-pub fn lookup_mock_server<R>(id: String, f: &Fn(&MockServer) -> R) -> Option<R> {
-    debug!("Looking up mock server with ID {}", id);
-    let map = MOCK_SERVERS.lock().unwrap();
-    match map.get(&id) {
-        Some(ref mock_server) => {
-            debug!("Found mock server, invoking function ...");
-            Some(f(mock_server))
-        },
-        None => {
-            debug!("Did not find mock server");
-            None
-        }
-    }
-}
-
-/// Looks up the mock server by port number, and passes it into the given closure. The result of the
-/// closure is returned wrapped in an `Option`. If no mock server is found with that port number, `None`
-/// is returned.
-pub fn lookup_mock_server_by_port<R>(mock_server_port: i32, f: &Fn(&MockServer) -> R) -> Option<R> {
-    let map = MOCK_SERVERS.lock().unwrap();
-    match map.iter().find(|ms| ms.1.port == mock_server_port ) {
-        Some(mock_server) => Some(f(mock_server.1)),
-        None => None
-    }
-}
-
-/// Iterates through all the mock servers, passing each one to the given closure.
-pub fn iterate_mock_servers(f: &mut FnMut(&String, &MockServer)) {
-    let map = MOCK_SERVERS.lock().unwrap();
-    for (key, value) in map.iter() {
-        f(key, value);
-    }
-}
-
-fn cleanup_mock_server_impl(mock_server: &mut MockServer) -> String {
-    mock_server.resources.clear();
-    if mock_server.server > 0 {
-        let server_raw = mock_server.server as *mut Listening;
-        let server_ref = unsafe { &mut *server_raw };
-        server_ref.close().unwrap();
-    }
-    mock_server.id.clone()
-}
-
-/// Shuts and cleans up the mock server with the given id. Returns true if a mock server was
-/// found, false otherwise.
-///
-/// **NOTE:** Although `close()` on the listerner for the mock server is called, this does not
-/// currently work and the listerner will continue handling requests. In this
-/// case, it will always return a 404 once the mock server has been cleaned up.
-pub fn shutdown_mock_server(id: &String) -> bool {
-    debug!("Shutting down mock server with ID {}", id);
-    let id_result = update_mock_server(id, &cleanup_mock_server_impl);
-
-    match id_result {
-        Some(ref id) => {
-            MOCK_SERVERS.lock().unwrap().remove(id);
-            true
-        },
-        None => false
-    }
-}
-
-/// Shuts and cleans up the mock server with the given port. Returns true if a mock server was
-/// found, false otherwise.
-///
-/// **NOTE:** Although `close()` on the listerner for the mock server is called, this does not
-/// currently work and the listerner will continue handling requests. In this
-/// case, it will always return a 404 once the mock server has been cleaned up.
-pub fn shutdown_mock_server_by_port(port: i32) -> bool {
-    debug!("Shutting down mock server with port {}", port);
-    let id_result = update_mock_server_by_port(port, &cleanup_mock_server_impl);
-
-    match id_result {
-        Some(ref id) => {
-            MOCK_SERVERS.lock().unwrap().remove(id);
-            true
-        },
-        None => false
-    }
-}
-
-/// Mock server errors
-pub enum MockServerError {
-  /// Invalid Pact Json
-  InvalidPactJson,
-  /// Failed to start the mock server
-  MockServerFailedToStart
+    MANAGER.lock().unwrap()
+        .get_or_insert_with(ServerManager::new)
+        .start_mock_server(id, pact, port as u16)
+        .map(|port| port as i32)
 }
 
 /// Creates a mock server. Requires the pact JSON as a string as well as the port for the mock
@@ -627,9 +240,12 @@ pub extern fn create_mock_server_ffi(pact_str: *const c_char, port: int32_t) -> 
 /// passed in, and if all requests have been matched, true is returned. False is returned if there
 /// is no mock server on the given port, or if any request has not been successfully matched.
 pub extern fn mock_server_matched(mock_server_port: i32) -> bool {
-  lookup_mock_server_by_port(mock_server_port, &|mock_server| {
-      mock_server.mismatches().is_empty()
-    }).unwrap_or(false)
+    MANAGER.lock().unwrap()
+        .get_or_insert_with(ServerManager::new)
+        .find_server_by_port_mut(mock_server_port as u16, &|mock_server| {
+            mock_server.mismatches().is_empty()
+        })
+        .unwrap_or(false)
 }
 
 /// External interface to check if a mock server has matched all its requests. The port number is
@@ -657,12 +273,14 @@ pub extern fn mock_server_matched_ffi(mock_server_port: int32_t) -> bool {
 /// If there is no mock server with the provided port number, `None` is returned.
 ///
 pub extern fn mock_server_mismatches(mock_server_port: i32) -> Option<std::string::String> {
-  lookup_mock_server_by_port(mock_server_port, &|mock_server| {
-    let mismatches = mock_server.mismatches().iter()
-      .map(|mismatch| mismatch.to_json() )
-      .collect::<Vec<serde_json::Value>>();
-    json!(mismatches).to_string()
-  })
+    MANAGER.lock().unwrap()
+        .get_or_insert_with(ServerManager::new)
+        .find_server_by_port_mut(mock_server_port as u16, &|mock_server| {
+            let mismatches = mock_server.mismatches().iter()
+            .map(|mismatch| mismatch.to_json() )
+            .collect::<Vec<serde_json::Value>>();
+            json!(mismatches).to_string()
+        })
 }
 
 /// External interface to get all the mismatches from a mock server. The port number of the mock
@@ -681,16 +299,18 @@ pub extern fn mock_server_mismatches(mock_server_port: i32) -> Option<std::strin
 #[no_mangle]
 pub extern fn mock_server_mismatches_ffi(mock_server_port: int32_t) -> *mut c_char {
     let result = catch_unwind(|| {
-        let result = update_mock_server_by_port(mock_server_port, &|ref mut mock_server| {
-            let mismatches = mock_server.mismatches().iter()
-                .map(|mismatch| mismatch.to_json() )
-                .collect::<Vec<serde_json::Value>>();
-            let json = json!(mismatches);
-            let s = CString::new(json.to_string()).unwrap();
-            let p = s.as_ptr();
-            mock_server.resources.push(s);
-            p
-        });
+        let result = MANAGER.lock().unwrap()
+            .get_or_insert_with(ServerManager::new)
+            .find_server_by_port_mut(mock_server_port as u16, &|ref mut mock_server| {
+                let mismatches = mock_server.mismatches().iter()
+                    .map(|mismatch| mismatch.to_json() )
+                    .collect::<Vec<serde_json::Value>>();
+                let json = json!(mismatches);
+                let s = CString::new(json.to_string()).unwrap();
+                let p = s.as_ptr();
+                mock_server.resources.push(s);
+                p
+            });
         match result {
             Some(p) => p as *mut _,
             None => 0 as *mut _
@@ -716,7 +336,9 @@ pub extern fn mock_server_mismatches_ffi(mock_server_port: int32_t) -> *mut c_ch
 #[no_mangle]
 pub extern fn cleanup_mock_server_ffi(mock_server_port: int32_t) -> bool {
     let result = catch_unwind(|| {
-        shutdown_mock_server_by_port(mock_server_port)
+        MANAGER.lock().unwrap()
+            .get_or_insert_with(ServerManager::new)
+            .shutdown_mock_server_by_port(mock_server_port as u16)
     });
 
     match result {
@@ -743,19 +365,23 @@ pub enum WritePactFileErr {
 /// Returns `Ok` if the pact file was successfully written. Returns an `Err` if the file can
 /// not be written, or there is no mock server running on that port.
 pub extern fn write_pact_file(mock_server_port: i32, directory: Option<String>) -> Result<(), WritePactFileErr> {
-    match lookup_mock_server_by_port(mock_server_port, &|mock_server| {
-        mock_server.write_pact(&directory)
-          .map(|_| ())
-          .map_err(|err| {
-            error!("Failed to write pact to file - {}", err);
-            WritePactFileErr::IOError
-          })
-    }) {
-      Some(result) => result,
-      None => {
-        error!("No mock server running on port {}", mock_server_port);
-        Err(WritePactFileErr::NoMockServer)
-      }
+    let opt_result = MANAGER.lock().unwrap()
+        .get_or_insert_with(ServerManager::new)
+        .find_server_by_port_mut(mock_server_port as u16, &|mock_server| {
+            mock_server.write_pact(&directory)
+                .map(|_| ())
+                .map_err(|err| {
+                    error!("Failed to write pact to file - {}", err);
+                    WritePactFileErr::IOError
+                })
+        });
+
+    match opt_result {
+        Some(result) => result,
+        None => {
+            error!("No mock server running on port {}", mock_server_port);
+            Err(WritePactFileErr::NoMockServer)
+        }
     }
 }
 
@@ -811,12 +437,8 @@ pub extern fn write_pact_file_ffi(mock_server_port: int32_t, directory: *const c
   }
 }
 
-#[cfg(test)]
-#[macro_use(expect)]
-extern crate expectest;
 
 #[cfg(test)]
-extern crate quickcheck;
-
-#[cfg(test)]
-mod tests;
+mod tests {
+    use super::*;
+}
