--- rust/pact_mock_server_async/Cargo.toml
@@ -1,37 +1,0 @@
-[package]
-name = "pact_mock_server_async"
-version = "0.4.1"
-authors = ["Ronald Holshausen <uglyog@gmail.com>"]
-description = "Pact-Rust support library that implements in process mock server"
-documentation = "https://docs.rs/pact_mock_server/0.4.1/pact_mock_server/"
-homepage = "http://www.pact.io"
-repository = "https://github.com/pact-foundation/pact-reference/blob/master/rust/pact_mock_server"
-readme = "README.md"
-keywords = ["testing", "pact", "cdc", "mockserver"]
-license = "MIT"
-exclude = [
-    "*.iml"
-]
-
-[dependencies]
-libc = "0.2.9"
-serde_json = "1.0"
-pact_matching = { version =  "0.5.0", path = "../pact_matching" }
-p-macro = "0.2.0"
-env_logger = "0.4.3"
-log = "0.3.8"
-maplit = "0.1.4"
-lazy_static = "0.2.8"
-uuid = { version = "0.6", features = ["v4"] }
-itertools = "0.6.1"
-tokio = "0.1.14"
-futures = "0.1"
-hyper = "0.12"
-bytes = "0.4.4"
-
-[lib]
-crate-type = ["cdylib", "staticlib", "rlib"]
-
-[dev-dependencies]
-quickcheck = "0.4.1"
-expectest = "0.9.1"
